test(db): split the smart-playlist suites under the module-size cap
pnpm lint:meta caps a Rust module at 400 code lines, and the three suites
grew past it. Split by job rather than by line count:

- smart_rules_analysis.rs takes the fields whose condition is not a plain
  comparison over a tracks column -- the analysis columns and their NULL, and
  last_played's scoped subquery.
- repo_smart_playlist_ordering.rs takes what order rows come back in and how
  many, leaving repo_smart_playlist_eval.rs owning which rows match.
- repo_smart_playlist_rules_column.rs takes the column's two shapes and their
  compatibility, leaving repo_smart_playlists.rs owning the CRUD.

support/rules.rs holds the four compiler fixtures both rule suites build on,
so they cannot drift. Every test moved unchanged; the counts are the same
29, 23 and 14.
